Slim down crossing example wrapper
